// src/types/crv-pools.ts
export interface CrvPool {
  id: string;
  address: string;
  coinsAddresses: string[];
  decimals: string[];
  virtualPrice: string;
  amplificationCoefficient: string;
  totalSupply: string;
  name: string;
  assetType: string;
  lpTokenAddress: string;
  priceOracle: null;
  priceOracles: null;
  symbol: string;
  implementation: string;
  assetTypeName: string;
  coins: Array<{
    address: string;
    usdPrice: number;
    decimals: string;
    isBasePoolLpToken: boolean;
    symbol: string;
    name: string;
    poolBalance: string;
  }>;
  poolUrls: { swap: string[]; deposit: string[]; withdraw: string[] };
  usdTotal: number;
  isMetaPool: boolean;
  usdTotalExcludingBasePool: number;
  gaugeAddress: string;
  gaugeRewards: any[];
  gaugeCrvApy: [number, number];
  gaugeFutureCrvApy: [number, number];
  usesRateOracle: boolean;
  isBroken: boolean;
  hasMethods: { exchange_received: boolean; exchange_extended: boolean };
  creationTs: number;
  creationBlockNumber: number;
  blockchainId: string;
  registryId: string;
}